Fix off-by-one error preventing ACG security from functioning
LS2Handle::SetAppId() stores the incoming servicePath from
jsservicelauncher/bootstrap* with a trailing "/" at the end.

LS2Handle::findMyAppId() searches for the servicePath with the trailing
"/" stripped.

This change fixes that.

Services with correct ACG configuration will no longer receive "The
service is not registered" errors from the Palmbus module, and so the
webos-service module will allow the ACG functionality to work.

Services without correct ACG configuration will receive "Invalid permissions
for (service id)", and will fall back to using the pre-ACG security. As
webos-service traps the "Invalid permissions" error, it will appear only
as "Deprecated security model is used. Please update configuration for ACG
security model."

Previous to this change, this same error would appear for all services,
regardless of if they had a correct ACG configuration or not.